Add routines for reading/writing bitmaps from/to disk
I'm not sure if this is the final solutions as it is quite messy
but for now it'll have to do.

One another way to do this would be interpret the fs block as a bitmap
where the first 8 bytes are bitmap length and rest of the block(s)
are bitmap bits. I might change the file system to use this technique
instead at some point.